Tracking corona virus total cases, deaths and new cases by country. Additionally, a detailed view is provided for the US(by state), Europe
This project is supported by a virtual team of open source collaborators that gather data, make visualizations and perform analysis regarding the COVID-19 epidemic. We saw that data scientists were struggling to share information with each other, so we used this project to aid in the sharing of data and information. This project aims to offer an accessible way for data professionals to share dashboards (that are updated with new data automatically) without requiring any expertise in front end development.